# scolaire-front/src/app/services/models/page-response-classe-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClasseResponse } from '../models/classe-response';
export interface PageResponseClasseResponse {
  content?: Array<ClasseResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}