# scolaire-front/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8092",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/halfYearly": {
      "get": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "getAllHalfYearly",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HalfYearlyResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "updateHalfYearly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HalfYearlyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HalfYearlyResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "newHalfYearly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HalfYearlyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HalfYearlyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "getAllCourses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "operationId": "updateCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "operationId": "addCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/classes": {
      "get": {
        "tags": [
          "Classes"
        ],
        "operationId": "getAllClasses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClasseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Classes"
        ],
        "operationId": "updateClasse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClasseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Classes"
        ],
        "operationId": "addClasse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClasseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/classes/{classeId}/assign-subject/{subjectId}": {
      "post": {
        "tags": [
          "Classes"
        ],
        "operationId": "assignSubjectToClasse",
        "parameters": [
          {
            "name": "classeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/halfYearly/{id}": {
      "get": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "getHalfYearlyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HalfYearlyResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "deleteHalfYearlyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/halfYearly/by-page": {
      "get": {
        "tags": [
          "HalfYearly"
        ],
        "operationId": "listAllHalfYearly",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseHalfYearlyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "getCourseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "operationId": "deleteCourseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/courses/by-page": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "listAllCourses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "tags": [
          "Classes"
        ],
        "operationId": "getClasseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Classes"
        ],
        "operationId": "deleteClasseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/classes/by-page": {
      "get": {
        "tags": [
          "Classes"
        ],
        "operationId": "listAllClasses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseClasseResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HalfYearlyRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "HalfYearlyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "CourseRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "subjectId": {
            "type": "integer",
            "format": "int64"
          },
          "halfYearlyId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "classeId",
          "name",
          "subjectId"
        ]
      },
      "CourseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "subjectId": {
            "type": "integer",
            "format": "int64"
          },
          "halfYearlyId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "ClasseRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "sectorId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "name"
        ]
      },
      "ClasseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "sectorId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageResponseHalfYearlyResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HalfYearlyResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseCourseResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseClasseResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClasseResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}